data: Simplify busconfig policy for tcmmd
Name ownership is denied by default, and since tcmmd does not export any
interfaces other than the three listed here, there is no point in
listing them separately in the policy.

If tcmmd does grow other interfaces in future which should not be
accessible to non-root processes, the access control for them should be
implemented in the daemon using polkit, rather than as D-Bus policy
which requires deep packet inspection. We’re trying to move away from
deep packet inspection because kdbus does not support it.