tcmmd: Re-add D-Bus invocation completion calls
These were erroneously removed in commit
bca5d716c36118718d9c4897cbc1e8fd288dad05, causing leaks. They were
removed because the methods are now marked as NoReply — but because
GDBus transfers ownership of the GDBusMethodInvocation object to the
service implementation when handling a method call, it __requires__ that
the service implementation call *_complete_*() exactly once on the
invocation object to free it.

This will cause warnings on the system bus due to sending unsolicited
replies to method calls — that’s
https://bugzilla.gnome.org/show_bug.cgi?id=755421.